+++ imagedl/modules/sources/__init__.py
@@ -3,6 +3,7 @@ from .base import BaseImageClient
 from .bing import BingImageClient
 from .i360 import I360ImageClient
 from .baidu import BaiduImageClient
+from .sogou import SogouImageClient
 from .google import GoogleImageClient
 from ..utils import BaseModuleBuilder
 from .yandex import YandexImageClient
@@ -15,7 +16,7 @@ class ImageClientBuilder(BaseModuleBuilder):
     REGISTERED_MODULES = {
         'BingImageClient': BingImageClient, 'BaiduImageClient': BaiduImageClient, 'GoogleImageClient': GoogleImageClient, 
         'I360ImageClient': I360ImageClient, 'PixabayImageClient': PixabayImageClient, 'YandexImageClient': YandexImageClient,
-        'DuckduckgoImageClient': DuckduckgoImageClient,
+        'DuckduckgoImageClient': DuckduckgoImageClient, 'SogouImageClient': SogouImageClient,
     }