+++ imagedl/modules/sources/baidu.py
@@ -18,13 +18,14 @@ class BaiduImageClient(BaseImageClient):
     source = 'BaiduImageClient'
     def __init__(self, **kwargs):
         super(BaiduImageClient, self).__init__(**kwargs)
-        self.default_headers = {
+        self.default_search_headers = {
             'Accept-Language': 'zh-CN,zh;q=0.8,zh-TW;q=0.7,zh-HK;q=0.5,en-US;q=0.3,en;q=0.2',
             'Connection': 'keep-alive',
             'Upgrade-Insecure-Requests': '1',
             'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/134.0.0.0 Safari/537.36',
         }
-        self.session.headers.update(self.default_headers)
+        self.default_headers = self.default_search_headers
+        self._initsession()
     '''_parsesearchresult'''
     def _parsesearchresult(self, search_result: str):
         # parse json text in safety