+++ imagedl/modules/sources/yandex.py
@@ -20,13 +20,14 @@ class YandexImageClient(BaseImageClient):
     def __init__(self, **kwargs):
         if 'maintain_session' not in kwargs: kwargs['maintain_session'] = True
         super(YandexImageClient, self).__init__(**kwargs)
-        self.default_headers = {
+        self.default_search_headers = {
             'User-Agent': 'Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/141.0.0.0 Safari/537.36',
             'Accept': 'text/html,application/xhtml+xml,application/xml;q=0.9,image/avif,image/webp,image/apng,*/*;q=0.8,application/signed-exchange;v=b3;q=0.7',
             'Accept-Encoding': 'gzip, deflate, br, zstd',
             'Accept-Language': 'zh-CN,zh;q=0.9,en-US;q=0.8,en;q=0.7',
         }
-        self.session.headers.update(self.default_headers)
+        self.default_headers = self.default_search_headers
+        self._initsession()
     '''_parsesearchresult'''
     def _parsesearchresult(self, search_result: str):
         # parse