+++ tests/test_image_clients.py
@@ -9,6 +9,7 @@ WeChat Official Account (微信公众号):
 from imagedl.modules.sources.bing import BingImageClient
 from imagedl.modules.sources.i360 import I360ImageClient
 from imagedl.modules.sources.baidu import BaiduImageClient
+from imagedl.modules.sources.sogou import SogouImageClient
 from imagedl.modules.sources.google import GoogleImageClient
 from imagedl.modules.sources.yandex import YandexImageClient
 from imagedl.modules.sources.pixabay import PixabayImageClient
@@ -24,6 +25,10 @@ image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)
 # baidu tests
 client = BaiduImageClient()
+image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
+client.download(image_infos, num_threadings=1)
+# sogou tests
+client = SogouImageClient()
 image_infos = client.search('Cute Dogs', search_limits=10, num_threadings=1)
 client.download(image_infos, num_threadings=1)
 # google tests