+++ docs/Changelog.md
@@ -1,5 +1,7 @@
 # Release Log
 
+- 2025-10-22: Released pyimagedl v0.1.7 — refactor codes for robust requests, fix requests bugs, and add sogou image search.
+
 - 2025-10-22: Released pyimagedl v0.1.6 — Fix serpapisearch bugs, more robust code structure, add DuckDuckGo image search.
 
 - 2025-10-21: Released pyimagedl v0.1.5 — Fix maintain session bugs, support pixabay and yandex image search.